socket.io

+++ src/reducers/UserReducer.js
@@ -86,14 +86,12 @@ const userSlice = createSlice({
     }).addCase(signUp.rejected,(state,{payload})=>{   
         state.error = payload.error
     }).addCase(getCurrentProfile.rejected,(state,{payload})=>{ 
-        if(payload.error){
+        if(payload && payload.error){
             state.loading = false
             state.signedIn = false
             state.currentProfile = null
             localStorage.clear()
         }  
-      
-        
     }).addCase(postStoryHistory.fulfilled,(state,{payload})=>{
         state.currentProfile = payload.profile
     }).addCase(postCollectionHistory.fulfilled,(state,{payload})=>{
@@ -104,13 +102,7 @@ const userSlice = createSlice({
        state.currentProfile = payload.profile
        state.loading = false
     })
-    // .addCase(fetchAllProfiles.fulfilled,(state,{ payload })=>{
-    //     state.profilesInView = payload.profileList
-    //     state.loading=false
-    // }).addCase(fetchAllProfiles.rejected,(state,{payload})=>{
-    //     state.error = payload.error
-    //     state.loading = false
-    // })
+ 
     .addCase(fetchProfile.pending,(state)=>{
         state.loading=true
     }).addCase(fetchProfile.fulfilled,(state,{ payload })=>{
